vpat 46: remaining aria-labels in preferences
- Linked remaining found inputs/menulists to their labels.
- Made the "Choose resolver" a proper visible
label, so that it does not oddly disappear if you click
on the dropdown with "Custom" showing after even if the
selection did not change. For the purpose of VPAT, a
visible label is always good for success criteria 3.3.2
https://www.w3.org/WAI/WCAG21/Understanding/labels-or-instructions
- Added explicit names to +/- buttons
- In cases when there is an input placed in the middle of a
contextual sentence (e.g. "Clear trash after <input> days"),
the label will announce both parts of the sentence, replacing
the <input> for "_".