vpat 22: add aria-disabled to editable-text on mac
VPAT review indicates that readonly editable text is not announced
as such by screen readers. This is true with voiceover and aria-disabled
is the only property that voiceover doesn't seem to ignore and it announces
such text as "dimmed". "aria-readonly", "readonly" are ignored.

I could not reproduce this on NVDA/JAWS - those announce readonly text as
expected, so if we don't need to add "aria-disabled", we shouldn't because
then text is announced as both readonly and disabled, which is confusing.

This addresses the reproduced issue on mac without affecting NVDA/JAWS.