vpat 50: rework data dir setting interface
Selecting custom data directory radio button in advanced
settings opens up a modal to select the directory. It is
not a typical behavior for radio button, so it conflicts
with 3.2.2 success criteria.

This is a rework of the data directory setting interface to be
similar to LABD. It does not have radio buttons and the
directory exporer pops up on button click