itembox focus edits and refactoring
- removed ztabindex logic from itemBox. It is no longer needed, adds
unnecessary complexity and is likely at the root of multiple glitches
if a plugin inserts an arbitrary row that does not have ztabindex set.
- if a creator row is deleted when the focus is inside of the row, focus
another creator row to not loose the focus.
- refactoring of hidden toolbarbuttons css so that the icons
are still hidden and don't occupy space (if desired) but are still visible
for screen readers, so they are focusable without JS changing their
visibility (this with ztabindex removal fixes vpat 24)
- removed escape_enter event from editable-text. It was a workaround to
know when itemBox should move focus back to itemTree. Unhandled Enter on an
input or Escape should focus itemTree (or reader) from wherever in the
itemPane/contextPane, so that logic is moved to itemDetails.js. To avoid conflicts
with Shift-Enter, do not propagate that event outside of multiline editable-text.
- removed not necessary keyboard nav handling from itemDetails.js. It was
only needed for mac, and is redundant if "Keyboard navigation" setting is on
- old handleKeyPress from itemBox is no longer relevant for most elements,
so it is removed and substituted with a dedicated handler just for creator row.
- moved the creator's paste handler into it's own dedicated function
from the autocomplete handler (which was confusing)
- itemBox can be not reachable in the duplicates pane if there are
a lot of records. The list of records will push itembox all the way below
the screen. So make it scrollable for now?