after new item creation, focus title in itemBox
When item is created manually, title field in itemBox
will be focused instead of the title in the itemPane as before.
If the itemBox is collapsed, it will be opened.

In addition, do not move the focus from the itemType menu
onto the title in the header if the item is created
via a keyboard shortcut and its type was immediately
changed. It is generally better to avoid moving focus on behalf
of the user, and now that the title is right below the
itemType, it's likely no longer necessary.

Added 'open' setter to item pane section for brevity