fixed focus handling with autocomplete in itemBox
If Enter keydown event reaches itemDetails, focus will go
to the itemTree (or, now, reader). Mozilla tries to stop
propagation of the Enter event on autocomplete fields
https://searchfox.org/mozilla-central/source/toolkit/content/widgets/autocomplete-input.js#564
which interferes with this event propagation. To avoid it,
capture Enter event on `editable-text` input, prevent default and
manually call input.handleEnter() to handle autocompletion. That way,
Enter on autocomplete will bubble to itemDetails as it should.

There is some odd behaviors with tagsBox from before this change but
still relevant here: Enter on last field will create a new editable-text
in the end as expected, hovewer Enter on any other autocomplete field
looses focus. This is to be handled next.