remove explicit tab handling from section header
It runs into issues when buttons get hidden or disabled
and the focus ends up being stuck not able to move past
some components.
(e.g. when viewing items in trash, tabbing will get
stuck on the "Notes" section header). It's best to just
remove it and let mozilla handle focus.

Previously added tab handling seemed to be necessary
only because "Keyboard navigation" on mac was disabled
so it's not needed per se.