improve JAWS in dialogs and non-main windows
JAWS expects a specific executable (firefox.exe) to apply
dialog-specific rules, which does not match in our case. It prevents
JAWS from deciding how to handle our dialogs and, for example,
it does not announce inputs correctly.

As a workaround, added `dialog=no` as a parameter to
some instances of opened dialogs or windows to force the HWND
window type be MozillaWindowClass which JAWS handles better.
This happens through Zotero.openWindow/Zotero.openDialog
where `dialog=no` is added only on windows. For now, only
do it for QF and dialogs opened in ZoteroPane.