tab semantics + combined item pane btns
- combine all itempane sidenav buttons into one focusable
group. Those buttons by themselves don't mean anything in
the context of keyboard navigation as they just scroll
to the section in the itemPane. In fact, having info,
abstract,attachments, etc. focusable and announceable by
screen readers is just confusing. However, we do want the group
of those buttons to be focusable to switch back to
zotero-context-pane-item-deck from zotero-context-pane-notes-deck
if the notes button in the sidenav is pressed.
- space/enter on the item pane buttons group will click
on the first (info) button. The info button also serves
as the aria-label
- added tablist, tab, tabpane roles semantics to itempane
buttons group and notes button to indicate that they switch
the views.