fix to stale trashed collections data
appearing in librariesCollectionsBox

- listen to 'collection' notifier events and re-render
the section if a relevant collection is trashed or modified.
That way, a trashed collection will be remoded, restored
collection will be added back and renaming a collection
will update the name in the librariesCollectionsBox.
- fixed a bug where a restored collection would not
be added into item._collections cache. Fixed false-positive
test for it.
- do not add trashed collections into items. _collections cache in
Zotero.Items._loadCollections. Otherwise, trashed collections will
appear in librariesCollectionsBox after the app is restarted.

Fixes: #4307