remove opacity when reference panel opens
Without the opacity on the reference panel, there is a shadow
appearing below and to the right of the panel which I don't think
we want. If we let the panel render with opacity:0.9 and only then
set opacity: 1, then the panel will no longer be transparent and
there will be no shadow.