vpat 7,14: popups count and cursor fixes
- vpat_7: when a popup is showing, mark menuseparators with
role="presentation" to clear whatever semantics a separator
has. It prevents the screen readers from counting it as
an interactable element while announcing the index of
a menuitem within the popup. It also makes the screen readers
count items across all sections, instead of stopping at
the first separator.
- - vpat_14: when menupopup is hiding, delete it and
immediately re-insert back into the DOM. It forces voiceover
to move cursor back to the focused element. Otherwise, it
is not aware that the popup went away, the cursor gets
stuck inside of the invisible menu, no elements
will be announced until the screen readers is reloaded.