vpat 25: idea to make prefs visible to jaws cursor
To help with JAWS refusing to sometimes interact with XUL
windows, open a window inside of a <browser> component
of an empty wrapper browser_wrapper.xhtml. That way, JAWS
can, for example, read all settings panes in cursor mode.

This has an additional benefit of fixing an issue that
prevents voiceover from announcing aria-live from
a XUL window as well. It's a known issue in firefox itself.
If we, say, load edit bibliography dialog inside of a browser,
we'll be able to announce instructions how to use it
onload, instead of adding an extra tabstop and changing
where focus lands on load (#4413). With this, we may be able to
add a shortcut to announce instructions how to use
which component regardless of where the focus is.

Finally, this approach may be easier to fix the issue of JAWS
not announcing inputs inside of dialogs (#4205) without
having to change 'dialog=yes/no' parameter and tweak the behavior
of how each dialog is opened.