handle opening wrapped browser win in one place
With ability to control if the <browser> should have
type="content" or not (needs to be set before the window
is loaded).
The main observed issue now is that setting type="content"
seems to be required for JAWS to read the content the
way we want with a cursor but it does create issues
by changing the appearance of default XUL components, such
as buttons, checkboxes, etc.
So if we do that, we may need to add specific non-native
styling for them...