add item-tree-menu-bar from itemTree.jsx
Instead of adding it to each .xhtml file and having to
remember to add it to every new window with itemTree.
If a menubar already exists, item-tree-menu-bar is not
added because it's fair to assume the other menubar
will add all needed menu options (as in zoteroPane.xhtml).

Revert earlier redundant changes to .xhtml files where
<item-tree-menu-bar> was added.