vpat 59: make quickFormat reference box separator text darker (#4400)
So that on Windows, the separator text on white background
has a high enough contrast ratio for accessibility

Also remove opacity when reference panel opens. Without the opacity
on the reference panel, there is a shadow appearing below and to the
right of the panel on Windows. If we let the panel render with
opacity: 0.9 and only then set opacity: 1, the panel will no longer
be transparent and there will be no shadow.