qf: remove transparent window background from win
Because it seems to still sometimes cause the qf window
to disappear. I could not reproduce it but disabling
background: transparent did help with the issue of qf
disappearing in the past. It allows us to remove very
specific chromemargins which, apparently, do not work
in all cases.

Removed border-radius from qf on windows because now that
the window is not transparent, its corners would be visible.
But the window itself has rounded corners so it doesn't look
that different.

The main drawback now is that a secondary window without titlebar
on win becomes resizable, and having background: transparent was
a workaround to disable that. So now, we set max width and height
so that the window cannot be stretched but, unfortunately, the
resize cursor appears when you hover over the edge of the dialog.