vpat 25/46: announce section headers of settings
- explicitly add aria-labels to groupbox and vbox
sections of the settings. It's required for voiceover
to announce that label when focus enters the group.
It's not needed for NVDA and JAWS as they recognize
most h2 labels as labels of grouppings.
- add all relevant additionl text for the section
as aria-description/aria-describedby so it is
announced after the section's name when focus enters it
- add role="group" to vbox sections, since vbox
does not have any semantics
- skip tabstop on scrollable "#prefs-content" container
since it is not required