vpat 14: aria-hidden instead of node removal
When menupopup is hiding, for a moment, mark it as
aria-hidden. It forces voiceover to move cursor back to
the previously focused element. Otherwise, it
is not aware that the popup went away, the cursor gets
stuck inside of the invisible menu, no elements
will be announced until the screen readers is reloaded.