qf: increase likeliness of dialog receiving focus
Tweak to Zotero.Utilities.Internal.activate to make
it more likely that qf receives focus. Addresses
the issue of qf not becoming focusable after the
first opening in word and makes it clearer when the
dialog appears in google docs when the dialog cannot
be pulled to the foreground.

getAttention() will flash the icon in the app bar
if the window could not be pulled into foreground,
for example with google docs.

focus() will try to bring the window to the foreground,
which fixes the issue of qf not getting focused
with word on the first opening. Also, without it,
a call to getAttention() interferes with focus landing
on the dialog during subsequent opening in gdocs.