server connector endpoint to cancel saving
- Added /connector/cancel endpoint to connector server that
the connector can call to cancel (or rather undo) the saving process if
it was started by mistake.
- If there is no session (meaning, saving did not start yet),
it will do nothing as the client will just need to not to hit the
/connector/save* endpoint.
- If there is a session and all saving is done, added items
will be erased right away.
- If there is a session and saving is not done, send all added
top-level items so far to the trash so they immediately go
away from the current view. Then, start a loop to check
periodically if the saving is done, and erase the items as
soon as it is. Letting all the saving process finish
seems like a more reliable way to achieve this without the risk of
corrupting the database.

A few additional fixes:
- fixed session not getting removed or garbage-collected
- fixed isSavingDone not returning true is any of the
attachments failed to get saved