fix lost focus due to whitespaces
Mainly, use trimmed creator field values to disregard
trailing whitespaces that are removed before saving
and do not cause a refresh.

- on shift-enter in unsaved creator row that only
has whitespace, do nothing, the focus will stay
in that row. It will be discarded if the focus leaves
it with nothing but whitespaces.
- fix focus getting lost on shift-enter on an editted
creator field after adding just some whitespaces.