minor refactoring
- both Choose and Use Custom Location buttons
- display input with default location after the
Restore to default location button, so that text
does not wrap or get cutoff if the path is long (but
can be scrolled).
- attemted to clear useDataDir pref in Zotero.DataDirectory.set
which lead to interesting breakage where restoring to
default directory would not pick the right pass by
having " default" appended to the end of default directory,
which does lead to wrong directory being used. Left
commented out for reference