context notes list focus handling followup
- remember the last focused note-row in the contextPane
and, if possible, refocus it on shift-tab from sidenav, on
tab from the reader into notes pane, or on tab into the
notes list from the collapsible-section. Otherwise, focus the search input.
Per: https://github.com/zotero/zotero/pull/4837#issuecomment-2478108186

- when a note is opened in the context pane, let shift-tab
from sidenave place focus on the links-box. That way,
it is not skipped during tab navigation.

- fix focus not entering note-editor on Tab from the "Go back"
button and instead wrapping around to the current tab