moving items to libraries/my publications via menu
- added My publications menuitem to item context menu to
add items to my publications without having to
drag-drop
- added Other Libraries menu to item context menu
to be able to add items to collections from other
libraries
- each group menu has a menuitem to create a new
collection in that group and add items there
- refactored addSelectedItemsToCollection to handle adding
items to collections outside of current library
- Refactor collectionTree and zoteroPane to pull out the
logic of verifying if an item can be added to a collection
or copied into a group to Zotero.Item. That way, both
collectionTree drag-drop API and zoteroPane contextmenus
can share Zotero.Item::canAddToContainer and
Zotero.Item::canAddToContainerAsync
- minor tweak to Zotero.Utilities.Internal.createMenuForTarget
to be able to force a menu (not menuitem) created library
even if there are not collections. Also, it now takes
an options object as a parameter, instead of multiple
positional parameters.