PageData actor: Settle for readyState "interactive"
Some pages seemingly never become "complete". "interactive" (which we
already use in TranslationChild) is equivalent to DOMContentLoaded, and
that's enough for everything PageDataChild needs to do right now.

This should make Scaffold test runs more reliable. TranslatorTester uses
{ requireSuccessfulStatus: true }, which calls the PageData actor, and
that was occasionally flaking and causing the test to time out.

Fixes #4869