unify collection canDropCheck with context menu
- move collectionTree canDropCheck and canDropCheckAsync
to Zotero.Collection::canMoveToTarget and Zotero.Collection::canMoveToTargetAsync
per https://github.com/zotero/zotero/pull/4420#issuecomment-2406579899
That way, we share the logic to determine if collection row can
be dropped onto another row, as well as to disable/enable context menus.
- allow to copy a collection to another library if
there already is a linked collection but it is in trash.
Fixes: #4862