make sidenav buttons individually focused
Instead of grouping the buttons and making them focusable
together. It seems like having the buttons groupped
ended up being more confusing then helpful per
https://forums.zotero.org/discussion/comment/481360/#Comment_481360.

In addition, now a keyboard-activated click on those buttons
will focus the header of the respective section, which can
actually be a useful shortcut for the navigation, especially
when there is a way to quickly place focus into the sidenav.

Fixes: #4930